backend/ml-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.13.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3 